backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tqdm==4.66.1
